Change entities structure, create a js file for every class in entities
