--- src/Scenes/Entities.js
@@ -1,460 +1,0 @@
-import 'phaser';
-import setScoreToStore from '../localStorage';
-
-class Entity extends Phaser.GameObjects.Sprite {
-  constructor(scene, x, y, key, type) {
-    super(scene, x, y, key);
-
-    this.scene = scene;
-    this.scene.add.existing(this);
-    this.scene.physics.world.enableBody(this, 0);
-    this.setData('type', type);
-    this.setData('isDead', false);
-    this.setData('score', 0);
-  }
-
-  explode(canDestroy) {
-    if (!this.getData('isDead')) {
-      this.setTexture('explosion');
-      this.play('explosion').setScale(0.4);
-
-      this.scene.sfx.explosions[Phaser.Math.Between(0, this.scene.sfx.explosions.length - 1)].play();
-
-      if (this.shootTimer !== undefined) {
-        if (this.shootTimer) {
-          this.shootTimer.remove(false);
-        }
-      }
-
-      this.setAngle(0);
-      this.body.setVelocity(0, 0);
-
-      this.on('animationcomplete', function () {
-        if (canDestroy) {
-          this.destroy();
-        } else {
-          this.setVisible(false);
-        }
-      }, this);
-
-      this.setData('isDead', true);
-    }
-  }
-}
-
-class Player extends Entity {
-  constructor(scene, x, y, key) {
-    super(scene, x, y, key, 'Player');
-
-    this.setData('speed', 200);
-    this.setData('isShooting', false);
-    this.setData('timerShootDelay', 10);
-    this.setData('timerShootTick', this.getData('timerShootDelay') - 1);
-    this.score = 0;
-  }
-
-  moveUp() {
-    this.body.velocity.y = -this.getData('speed');
-  }
-
-  moveDown() {
-    this.body.velocity.y = this.getData('speed');
-  }
-
-  moveLeft() {
-    this.body.velocity.x = -this.getData('speed');
-  }
-
-  moveRight() {
-    this.body.velocity.x = this.getData('speed');
-  }
-
-  updateScore(enemy) {
-    if (enemy.getData('type') === 'CarrierShip') {
-      this.setData('score', this.getData('score') + 25);
-    } else if (enemy.getData('type') === 'GunShip') {
-      this.setData('score', this.getData('score') + 15);
-    } else if (enemy.getData('type') === 'ChaserShip') {
-      this.setData('score', this.getData('score') + 10);
-    } else {
-      this.setData('score', this.getData('score') + 5);
-    }
-  }
-
-  updateScoretoLocal() {
-    setScoreToStore(this.getData('score'));
-  }
-
-  update() {
-    this.body.setVelocity(0, 0);
-
-    this.x = Phaser.Math.Clamp(this.x, 0, this.scene.game.config.width);
-    this.y = Phaser.Math.Clamp(this.y, 0, this.scene.game.config.height);
-
-    if (this.getData('isShooting')) {
-      if (this.getData('timerShootTick') < this.getData('timerShootDelay')) {
-        this.setData('timerShootTick', this.getData('timerShootTick') + 1);
-      } else {
-        const laser = new PlayerLaser(this.scene, this.x, this.y);
-        this.scene.playerLasers.add(laser);
-
-        this.scene.sfx.laser.play();
-        this.setData('timerShootTick', 0);
-      }
-    }
-  }
-
-  onDestroy() {
-    this.scene.time.addEvent({ // go to game over scene
-      delay: 500,
-      callback() {
-        this.scene.scene.start('GameOver');
-      },
-      callbackScope: this,
-      loop: false,
-    });
-  }
-}
-
-class ChaserShip extends Entity {
-  constructor(scene, x, y) {
-    super(scene, x, y, 'enemy', 'ChaserShip');
-    this.body.velocity.y = Phaser.Math.Between(50, 100);
-    this.shootTimer = this.scene.time.addEvent({
-      delay: 1000,
-      callback() {
-        const laser = new EnemyLaser(
-          this.scene,
-          this.x,
-          this.y,
-        );
-        laser.setScale(this.scaleX);
-        this.scene.enemyLasers.add(laser);
-      },
-      callbackScope: this,
-      loop: true,
-    });
-    this.states = {
-      MOVE_DOWN: 'MOVE_DOWN',
-      CHASE: 'CHASE',
-    };
-    this.state = this.states.MOVE_DOWN;
-  }
-
-  onDestroy() {
-    if (this.shootTimer !== undefined) {
-      if (this.shootTimer) {
-        this.shootTimer.remove(false);
-      }
-    }
-  }
-
-  update() {
-    if (!this.getData('isDead') && this.scene.player) {
-      if (Phaser.Math.Distance.Between(
-        this.x,
-        this.y,
-        this.scene.player.x,
-        this.scene.player.y,
-      ) < 320) {
-        this.state = this.states.CHASE;
-      }
-
-      if (this.state == this.states.CHASE) {
-        const dx = this.scene.player.x - this.x;
-        const dy = this.scene.player.y - this.y;
-
-        const angle = Math.atan2(dy, dx);
-
-        const speed = 100;
-        this.body.setVelocity(
-          Math.cos(angle) * speed,
-          Math.sin(angle) * speed,
-        );
-      }
-
-      if (this.x < this.scene.player.x) {
-        this.angle -= 5;
-      } else {
-        this.angle += 5;
-      }
-
-      if (enemy.x < -enemy.displayWidth
-        || enemy.x > this.game.config.width + enemy.displayWidth
-        || enemy.y < -enemy.displayHeight * 4
-        || enemy.y > this.game.config.height + enemy.displayHeight) {
-        if (enemy) {
-          if (enemy.onDestroy !== undefined) {
-            enemy.explode();
-          }
-        }
-      }
-
-      for (var i = 0; i < this.enemyLasers.getChildren().length; i++) {
-        var laser = this.enemyLasers.getChildren()[i];
-        laser.update();
-
-        if (laser.x < -laser.displayWidth
-          || laser.x > this.game.config.width + laser.displayWidth
-          || laser.y < -laser.displayHeight * 4
-          || laser.y > this.game.config.height + laser.displayHeight) {
-          if (laser) {
-            laser.destroy();
-          }
-        }
-      }
-
-      for (var i = 0; i < this.playerLasers.getChildren().length; i++) {
-        var laser = this.playerLasers.getChildren()[i];
-        laser.update();
-
-        if (laser.x < -laser.displayWidth
-          || laser.x > this.game.config.width + laser.displayWidth
-          || laser.y < -laser.displayHeight * 4
-          || laser.y > this.game.config.height + laser.displayHeight) {
-          if (laser) {
-            laser.destroy();
-          }
-        }
-      }
-    }
-  }
-}
-
-class CarrierShip extends Entity {
-  constructor(scene, x, y) {
-    super(scene, x, y, 'enemy1', 'CarrierShip');
-    this.body.velocity.y = Phaser.Math.Between(50, 100);
-    this.shootTimer = this.scene.time.addEvent({
-      delay: 2000,
-      callback() {
-        const laser = new EnemyLaser(
-          this.scene,
-          this.x,
-          this.y,
-        );
-        laser.setScale(this.scaleX);
-        this.scene.enemyLasers.add(laser);
-      },
-      callbackScope: this,
-      loop: true,
-    });
-    this.states = {
-      MOVE_DOWN: 'MOVE_DOWN',
-      CHASE: 'CHASE',
-    };
-    this.state = this.states.MOVE_DOWN;
-  }
-
-  onDestroy() {
-    if (this.shootTimer !== undefined) {
-      if (this.shootTimer) {
-        this.shootTimer.remove(false);
-      }
-    }
-  }
-
-  update() {
-    if (!this.getData('isDead') && this.scene.player) {
-      if (Phaser.Math.Distance.Between(
-        this.x,
-        this.y,
-        this.scene.player.x,
-        this.scene.player.y,
-      ) < 320) {
-        this.state = this.states.CHASE;
-      }
-
-      if (this.state == this.states.CHASE) {
-        const dx = this.scene.player.x - this.x;
-        const dy = this.scene.player.y - this.y;
-
-        const angle = Math.atan2(dy, dx);
-
-        const speed = 100;
-        this.body.setVelocity(
-          Math.cos(angle) * speed,
-          Math.sin(angle) * speed,
-        );
-      }
-    }
-  }
-}
-
-class GunShip extends Entity {
-  constructor(scene, x, y) {
-    super(scene, x, y, 'enemy2', GunShip);
-    this.body.velocity.y = Phaser.Math.Between(50, 100);
-    this.shootTimer = this.scene.time.addEvent({
-      delay: 4000,
-      callback() {
-        const laser = new EnemyLaser(
-          this.scene,
-          this.x,
-          this.y,
-        );
-        laser.setScale(this.scaleX);
-        this.scene.enemyLasers.add(laser);
-      },
-      callbackScope: this,
-      loop: true,
-    });
-    this.states = {
-      MOVE_DOWN: 'MOVE_DOWN',
-      CHASE: 'CHASE',
-    };
-    this.state = this.states.MOVE_DOWN;
-  }
-
-  onDestroy() {
-    if (this.shootTimer !== undefined) {
-      if (this.shootTimer) {
-        this.shootTimer.remove(false);
-      }
-    }
-  }
-
-  update() {
-    if (!this.getData('isDead') && this.scene.player) {
-      if (Phaser.Math.Distance.Between(
-        this.x,
-        this.y,
-        this.scene.player.x,
-        this.scene.player.y,
-      ) < 320) {
-        this.state = this.states.CHASE;
-      }
-
-      if (this.state == this.states.CHASE) {
-        const dx = this.scene.player.x - this.x;
-        const dy = this.scene.player.y - this.y;
-
-        const angle = Math.atan2(dy, dx);
-
-        const speed = 100;
-        this.body.setVelocity(
-          Math.cos(angle) * speed,
-          Math.sin(angle) * speed,
-        );
-      }
-    }
-  }
-}
-
-class HollowShip extends Entity {
-  constructor(scene, x, y) {
-    super(scene, x, y, 'enemy3', 'HollowShip');
-    this.body.velocity.y = Phaser.Math.Between(50, 100);
-    this.shootTimer = this.scene.time.addEvent({
-      delay: 3000,
-      callback() {
-        const laser = new EnemyLaser(
-          this.scene,
-          this.x,
-          this.y,
-        );
-        laser.setScale(this.scaleX);
-        this.scene.enemyLasers.add(laser);
-      },
-      callbackScope: this,
-      loop: true,
-    });
-    this.states = {
-      MOVE_DOWN: 'MOVE_DOWN',
-      CHASE: 'CHASE',
-    };
-    this.state = this.states.MOVE_DOWN;
-  }
-
-  onDestroy() {
-    if (this.shootTimer !== undefined) {
-      if (this.shootTimer) {
-        this.shootTimer.remove(false);
-      }
-    }
-  }
-
-  update() {
-    if (!this.getData('isDead') && this.scene.player) {
-      if (Phaser.Math.Distance.Between(
-        this.x,
-        this.y,
-        this.scene.player.x,
-        this.scene.player.y,
-      ) < 320) {
-        this.state = this.states.CHASE;
-      }
-
-      if (this.state == this.states.CHASE) {
-        const dx = this.scene.player.x - this.x;
-        const dy = this.scene.player.y - this.y;
-
-        const angle = Math.atan2(dy, dx);
-
-        const speed = 100;
-        this.body.setVelocity(
-          Math.cos(angle) * speed,
-          Math.sin(angle) * speed,
-        );
-      }
-
-      if (this.x < this.scene.player.x) {
-        this.angle -= 5;
-      } else {
-        this.angle += 5;
-      }
-    }
-  }
-}
-
-class EnemyLaser extends Entity {
-  constructor(scene, x, y) {
-    super(scene, x, y, 'laserShooter');
-    this.body.velocity.y = 200;
-  }
-}
-
-class PlayerLaser extends Entity {
-  constructor(scene, x, y) {
-    super(scene, x, y, 'laserShooter');
-    this.body.velocity.y = -200;
-  }
-}
-
-class ScrollingBackground {
-  constructor(scene, key, velocityY) {
-    this.scene = scene;
-    this.key = key;
-    this.velocityY = velocityY;
-    this.layers = this.scene.add.group();
-    this.createLayers();
-  }
-
-  createLayers() {
-    for (let i = 0; i < 2; i++) {
-      const layer = this.scene.add.sprite(0, 0, this.key);
-      layer.y = (layer.displayHeight * i);
-      const flipX = Phaser.Math.Between(0, 10) >= 5 ? -1 : 1;
-      const flipY = Phaser.Math.Between(0, 10) >= 5 ? -1 : 1;
-      layer.setScale(flipX * 2, flipY * 2);
-      layer.setDepth(-5 - (i - 1));
-      this.scene.physics.world.enableBody(layer, 0);
-      layer.body.velocity.y = this.velocityY;
-
-      this.layers.add(layer);
-    }
-  }
-
-  update() {
-    if (this.layers.getChildren()[0].y > 0) {
-      for (let i = 0; i < this.layers.getChildren().length; i++) {
-        const layer = this.layers.getChildren()[i];
-        layer.y = (-layer.displayHeight) + (layer.displayHeight * i);
-      }
-    }
-  }
-}
-
-export {
-  Player, ChaserShip, CarrierShip, GunShip, HollowShip, EnemyLaser, PlayerLaser, ScrollingBackground,
-};
