+++ src/Scenes/entities/Gunship.js
@@ -1,0 +1,63 @@
+import Phaser from 'phaser';
+import Entity from './Entities';
+import EnemyLaser from './Enemylaser';
+
+export default class GunShip extends Entity {
+    constructor(scene, x, y) {
+      super(scene, x, y, 'enemy2', GunShip);
+      this.body.velocity.y = Phaser.Math.Between(50, 100);
+      this.shootTimer = this.scene.time.addEvent({
+        delay: 4000,
+        callback() {
+          const laser = new EnemyLaser(
+            this.scene,
+            this.x,
+            this.y,
+          );
+          laser.setScale(this.scaleX);
+          this.scene.enemyLasers.add(laser);
+        },
+        callbackScope: this,
+        loop: true,
+      });
+      this.states = {
+        MOVE_DOWN: 'MOVE_DOWN',
+        CHASE: 'CHASE',
+      };
+      this.state = this.states.MOVE_DOWN;
+    }
+  
+    onDestroy() {
+      if (this.shootTimer !== undefined) {
+        if (this.shootTimer) {
+          this.shootTimer.remove(false);
+        }
+      }
+    }
+  
+    update() {
+      if (!this.getData('isDead') && this.scene.player) {
+        if (Phaser.Math.Distance.Between(
+          this.x,
+          this.y,
+          this.scene.player.x,
+          this.scene.player.y,
+        ) < 320) {
+          this.state = this.states.CHASE;
+        }
+  
+        if (this.state == this.states.CHASE) {
+          const dx = this.scene.player.x - this.x;
+          const dy = this.scene.player.y - this.y;
+  
+          const angle = Math.atan2(dy, dx);
+  
+          const speed = 100;
+          this.body.setVelocity(
+            Math.cos(angle) * speed,
+            Math.sin(angle) * speed,
+          );
+        }
+      }
+    }
+  }
